backend/requirements.txt
tqdm
torch==2.4.1
torchaudio==2.4.1
opencv-python
ffmpegcv
numpy
python_speech_features
scipy
scenedetect
scikit-learn
gdown
torchvision==0.19.1
pandas
transformers
accelerate
datasets
google-genai
pysubs2
boto3
fastapi[standard]
whisperx
pyannote.audio>=3.3.2
matplotlib